renamed the finufft dir
