+++ .gitignore
@@ -1,17 +1,14 @@
 spread1d
 spread2d
-compare
-finufft2d_test
+spread3d
 interp2d
+interp3d
 cufinufft2d1_test
 cufinufft2d1many_test
 cufinufft2d2_test
 cufinufft2d2many_test
-spreadinterp_test
-benchmark/cunfft_timing
-benchmark/spreadtestnd_mel
-benchmark/nfft_simpletest
-benchmark/result/*.out
+cufinufft3d1_test
+cufinufft3d2_test
 *.o
 __pycache__
 lib
