+++ contrib/common.h
