+++ contrib/ker_horner_allw_loop.c
