+++ contrib/ker_lowupsampfac_horner_allw_loop.c
