+++ contrib/legendre_rule_fast.c
