+++ contrib/legendre_rule_fast.h
