+++ contrib/spreadinterp.cpp
