+++ contrib/utils.cpp
