+++ contrib/utils.h
