+++ test/cufinufft3d1_test.cu
@@ -6,7 +6,7 @@
 
 #include "../src/spreadinterp.h"
 #include "../src/cufinufft.h"
-#include "../finufft/utils.h"
+#include "../contrib/utils.h"
 
 using namespace std;
