+++ test/spread_2d.cu
@@ -6,7 +6,7 @@
 #include <algorithm>
 #include "../src/cufinufft.h"
 #include "../src/spreadinterp.h"
-#include "../finufft/utils.h"
+#include "../contrib/utils.h"
 
 using namespace std;
